--- App/AppStore.swift
@@ -1,8 +1,0 @@
-//
-//  AppStore.swift
-//  GlucoseDirect
-//
-
-import Foundation
-
-typealias AppStore = Store<AppState, AppAction>
