+++ App/Modules/GlucoseNotification/GlucoseNotification.swift
@@ -8,13 +8,13 @@ import Foundation
 import UIKit
 import UserNotifications
 
-func glucoseNotificationMiddelware() -> Middleware<AppState, AppAction> {
+func glucoseNotificationMiddelware() -> Middleware<DirectState, DirectAction> {
     return glucoseNotificationMiddelware(service: LazyService<GlucoseNotificationService>(initialization: {
         GlucoseNotificationService()
     }))
 }
 
-private func glucoseNotificationMiddelware(service: LazyService<GlucoseNotificationService>) -> Middleware<AppState, AppAction> {
+private func glucoseNotificationMiddelware(service: LazyService<GlucoseNotificationService>) -> Middleware<DirectState, DirectAction> {
     return { state, action, _ in
         switch action {
         case .setHighGlucoseAlarmSound(sound: let sound):
@@ -117,7 +117,7 @@ private class GlucoseNotificationService {
     }
 
     func setGlucoseBadge(glucose: Glucose, glucoseUnit: GlucoseUnit) {
-        NotificationService.shared.ensureCanSendNotification { state in
+        DirectNotifications.shared.ensureCanSendNotification { state in
             DirectLog.info("Glucose info, state: \(state)")
 
             guard state != .none else {
@@ -135,7 +135,7 @@ private class GlucoseNotificationService {
     }
 
     func setGlucoseNotification(glucose: Glucose, glucoseUnit: GlucoseUnit) {
-        NotificationService.shared.ensureCanSendNotification { state in
+        DirectNotifications.shared.ensureCanSendNotification { state in
             DirectLog.info("Glucose info, state: \(state)")
 
             guard state != .none else {
@@ -162,12 +162,12 @@ private class GlucoseNotificationService {
                                        glucose.minuteChange?.asMinuteChange(glucoseUnit: glucoseUnit) ?? "?"
             )
 
-            NotificationService.shared.add(identifier: Identifier.sensorGlucoseAlarm.rawValue, content: notification)
+            DirectNotifications.shared.add(identifier: Identifier.sensorGlucoseAlarm.rawValue, content: notification)
         }
     }
 
     func setLowGlucoseAlarm(glucose: Glucose, glucoseUnit: GlucoseUnit, ignoreMute: Bool, sound: NotificationSound) {
-        NotificationService.shared.ensureCanSendNotification { state in
+        DirectNotifications.shared.ensureCanSendNotification { state in
             DirectLog.info("Glucose alert, state: \(state)")
 
             guard state != .none else {
@@ -195,16 +195,16 @@ private class GlucoseNotificationService {
                                        glucose.minuteChange?.asMinuteChange(glucoseUnit: glucoseUnit) ?? "?"
             )
 
-            NotificationService.shared.add(identifier: Identifier.sensorGlucoseAlarm.rawValue, content: notification)
+            DirectNotifications.shared.add(identifier: Identifier.sensorGlucoseAlarm.rawValue, content: notification)
 
             if state == .sound {
-                NotificationService.shared.playSound(ignoreMute: ignoreMute, sound: sound)
+                DirectNotifications.shared.playSound(ignoreMute: ignoreMute, sound: sound)
             }
         }
     }
 
     func setHighGlucoseAlarm(glucose: Glucose, glucoseUnit: GlucoseUnit, ignoreMute: Bool, sound: NotificationSound) {
-        NotificationService.shared.ensureCanSendNotification { state in
+        DirectNotifications.shared.ensureCanSendNotification { state in
             DirectLog.info("Glucose alert, state: \(state)")
 
             guard state != .none else {
@@ -232,10 +232,10 @@ private class GlucoseNotificationService {
                                        glucose.minuteChange?.asMinuteChange(glucoseUnit: glucoseUnit) ?? "?"
             )
 
-            NotificationService.shared.add(identifier: Identifier.sensorGlucoseAlarm.rawValue, content: notification)
+            DirectNotifications.shared.add(identifier: Identifier.sensorGlucoseAlarm.rawValue, content: notification)
 
             if state == .sound {
-                NotificationService.shared.playSound(ignoreMute: ignoreMute, sound: sound)
+                DirectNotifications.shared.playSound(ignoreMute: ignoreMute, sound: sound)
             }
         }
     }
