+++ App/Modules/ScreenLock/ScreenLock.swift
@@ -1,0 +1,24 @@
+//
+//  ScreenLock.swift
+//  GlucoseDirect
+//
+
+import Combine
+import Foundation
+import UIKit
+
+func screenLockMiddleware() -> Middleware<DirectState, DirectAction> {
+    return { _, action, _ in
+        DirectLog.info("Triggered action: \(action)")
+
+        switch action {
+        case .setPreventScreenLock(enabled: let enabled):
+            UIApplication.shared.isIdleTimerDisabled = enabled
+
+        default:
+            break
+        }
+
+        return Empty().eraseToAnyPublisher()
+    }
+}
