+++ App/Modules/SensorConnector/BubbleConnection/BubbleConnection.swift
@@ -9,10 +9,10 @@ import Foundation
 
 // MARK: - BubbleConnection
 
-class BubbleConnection: SensorBLEConnectionBase, IsTransmitter {
+class BubbleConnection: SensorBluetoothConnection, IsTransmitter {
     // MARK: Lifecycle
 
-    init(subject: PassthroughSubject<AppAction, AppError>) {
+    init(subject: PassthroughSubject<DirectAction, AppError>) {
         DirectLog.info("init")
         super.init(subject: subject, serviceUUID: CBUUID(string: "6E400001-B5A3-F393-E0A9-E50E24DCCA9E"))
     }
@@ -125,7 +125,7 @@ class BubbleConnection: SensorBLEConnectionBase, IsTransmitter {
                     resetBuffer()
                     return
                 }
-                
+
                 guard let fram = type == .libre1
                     ? rxBuffer[..<expectedBufferSize]
                     : Libre2EUtility.decryptFRAM(uuid: uuid, patchInfo: patchInfo, fram: rxBuffer[..<expectedBufferSize])
@@ -142,7 +142,7 @@ class BubbleConnection: SensorBLEConnectionBase, IsTransmitter {
 
                 } else if sensor.age > sensor.warmupTime {
                     let readings = LibreUtility.parseFRAM(calibration: sensor.factoryCalibration, pairingTimestamp: sensor.pairingTimestamp, fram: fram)
-                    
+
                     sendUpdate(age: sensor.age, state: sensor.state)
                     sendUpdate(sensorSerial: sensor.serial ?? "", readings: readings.history + readings.trend)
                 } else if sensor.age <= sensor.warmupTime {
