+++ App/Modules/SensorConnector/Libre2Connection/Libre2Pairing.swift
@@ -16,7 +16,7 @@ import CoreNFC
 class Libre2Pairing: NSObject, NFCTagReaderSessionDelegate {
     // MARK: Lifecycle
 
-    init(subject: PassthroughSubject<AppAction, AppError>) {
+    init(subject: PassthroughSubject<DirectAction, AppError>) {
         self.subject = subject
     }
 
@@ -203,7 +203,7 @@ class Libre2Pairing: NSObject, NFCTagReaderSessionDelegate {
     // MARK: Private
 
     private var session: NFCTagReaderSession?
-    private weak var subject: PassthroughSubject<AppAction, AppError>?
+    private weak var subject: PassthroughSubject<DirectAction, AppError>?
 
     private let nfcQueue = DispatchQueue(label: "libre-direct.nfc-queue")
     private let unlockCode: UInt32 = 42
@@ -281,7 +281,7 @@ private enum Subcommand: UInt8, CustomStringConvertible {
 class Libre2Pairing: NSObject {
     // MARK: Lifecycle
 
-    init(subject: PassthroughSubject<AppAction, AppError>) {}
+    init(subject: PassthroughSubject<DirectAction, AppError>) {}
 
     // MARK: Internal
