+++ App/Views/Overview/ConnectionView.swift
@@ -8,7 +8,7 @@ import SwiftUI
 // MARK: - ConnectionView
 
 struct ConnectionView: View {
-    @EnvironmentObject var store: AppStore
+    @EnvironmentObject var store: DirectStore
 
     var body: some View {
         Section(
