+++ App/Views/Overview/GlucoseView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 struct GlucoseView: View {
     // MARK: Internal
 
-    @EnvironmentObject var store: AppStore
+    @EnvironmentObject var store: DirectStore
 
     var body: some View {
         if let latestGlucose = store.state.latestGlucose {
