+++ App/Views/Overview/SensorView.swift
@@ -8,7 +8,7 @@ import SwiftUI
 // MARK: - SensorView
 
 struct SensorView: View {
-    @EnvironmentObject var store: AppStore
+    @EnvironmentObject var store: DirectStore
 
     var body: some View {
         Group {
