+++ App/Views/Settings/AppleExportSettingsView.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct AppleExportSettingsView: View {
     // MARK: Internal
 
-    @EnvironmentObject var store: AppStore
+    @EnvironmentObject var store: DirectStore
 
     var body: some View {
         Section(
