+++ App/Views/SettingsView.swift
@@ -6,7 +6,7 @@
 import SwiftUI
 
 struct SettingsView: View {
-    @EnvironmentObject var store: AppStore
+    @EnvironmentObject var store: DirectStore
 
     var body: some View {
         VStack {
