+++ Library/DirectStore.swift
@@ -1,0 +1,8 @@
+//
+//  DirectStore.swift
+//  GlucoseDirect
+//
+
+import Foundation
+
+typealias DirectStore = Store<DirectState, DirectAction>
